Use references everywhere in vehicle code

+++ TombEngine/Objects/TR4/Vehicles/jeep_info.h
@@ -11,21 +11,21 @@ namespace TEN::Entities::Vehicles
 		short ExtraRotation		 = 0;
 		short ExtraRotationDrift = 0;
 
-		int Revs = 0;
-		short EngineRevs = 0;
-		int Pitch = 0;
-		int CameraElevation = 0;
-		short Gear = 0;
-		short TrackMesh = 0;
+		int	  Revs			  = 0;
+		short EngineRevs	  = 0;
+		int	  Pitch			  = 0;
+		int	  CameraElevation = 0;
+		short Gear			  = 0;
+		short TrackMesh		  = 0;
 
 		short FrontRightWheelRotation = 0;
 		short FrontLeftWheelRotation  = 0;
 		short BackRightWheelRotation  = 0;
 		short BackLeftWheelRotation   = 0;
 
-		char SmokeStart = 0;
+		char SmokeStart		 = 0;
 		bool DisableDismount = false;
 
-		short Flags = 0;
+		int Flags = 0;
 	};
 }
