+++ TombEngine/Objects/TR4/Vehicles/motorbike_info.h
@@ -6,21 +6,21 @@ namespace TEN::Entities::Vehicles
 	{
 		int Velocity = 0;
 
-		short TurnRate = 0;
-		short MomentumAngle = 0;
-		short ExtraRotation = 0;
+		short TurnRate			= 0;
+		short MomentumAngle		= 0;
+		short ExtraRotation		= 0;
 		short WallShiftRotation = 0;
 
-		int Revs = 0;
-		int EngineRevs = 0;
-		int Pitch = 0;
-		short LightPower = 0;
-		short LeftWheelRotation = 0;
+		int	  Revs				  = 0;
+		int	  EngineRevs		  = 0;
+		int	  Pitch				  = 0;
+		short LightPower		  = 0;
+		short LeftWheelRotation	  = 0;
 		short RightWheelsRotation = 0;
 
-		char ExhaustStart = 0;
+		char ExhaustStart	 = 0;
 		bool DisableDismount = false;
 
-		short Flags = 0;
+		int Flags = 0;
 	};
 }
